# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: IntelliDesk API
  version: 0.1.0
  description: |
    与 PRD_Backend_Detailed 对齐。鉴权：请求头 `X-Tenant-Id`（必填）、`X-User-Id`（可选，演示用已存在用户 id）。
servers:
  - url: http://localhost:4000
    description: 本地后端直连（默认 PORT）
  - url: http://localhost:4001
    description: 经前端 Vite 同源代理（npm run dev:api，浏览器站点根）
...
components:
  securitySchemes:
    TenantHeader:
      type: apiKey
      in: header
      name: X-Tenant-Id
  parameters:
    TenantId:
      name: X-Tenant-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid
...
security:
  - TenantHeader: []
paths:
...
  /api/settings/agent-seats:
    get:
      tags: [settings]
      parameters:
        - $ref: '#/components/parameters/TenantId'
    post:
      tags: [settings]
      parameters:
        - $ref: '#/components/parameters/TenantId'
  /api/settings/agent-seats/{id}:
    patch:
      tags: [settings]
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
    delete:
      tags: [settings]
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
  /api/settings/sla-rules:
    get:
      tags: [settings]
      parameters:
        - $ref: '#/components/parameters/TenantId'
    post:
      tags: [settings]
      parameters:
        - $ref: '#/components/parameters/TenantId'
  /api/settings/sla-rules/{id}:
    patch:
      tags: [settings]
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
    delete:
      tags: [settings]
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
  /api/settings/auto-reply-rules:
    get:
      tags: [settings]
      parameters:
        - $ref: '#/components/parameters/TenantId'
    post:
      tags: [settings]
      parameters:
        - $ref: '#/components/parameters/TenantId'
  /api/settings/auto-reply-rules/{id}:
    patch:
      tags: [settings]
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
    delete:
      tags: [settings]
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
  /api/settings/routing-rule-options:
    get:
      summary: 分配规则下拉数据（平台类型去重、渠道列表、售后类型字典）
      tags: [settings]
      parameters:
        - $ref: '#/components/parameters/TenantId'
  /api/settings/ticket-routing-rules:
    get:
      tags: [settings]
      parameters:
        - $ref: '#/components/parameters/TenantId'
    post:
      tags: [settings]
      parameters:
        - $ref: '#/components/parameters/TenantId'
  /api/settings/ticket-routing-rules/{id}:
    patch:
      tags: [settings]
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
    delete:
      tags: [settings]
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }